feat: Add support for Gmail labels in GMAIL_SCOPES
This commit adds the necessary scope https://www.googleapis.com/auth/gmail.labels to the GMAIL_SCOPES list in the settings.py file. This scope allows the application to access and modify Gmail labels. This change enables the application to perform actions related to Gmail labels, such as moving emails to specific mailboxes.