feat: Add unit test for TestAutomate class
This commit adds a new unit test for the TestAutomate class in the test_automate.py file. The test_automate.py file is located in the tests directory. The new test_automate.py file contains a single test method named test_automate, which currently does nothing. This test will be used to verify the functionality of the TestAutomate class in future development.